vdf: merge VDF impl into `crypto/` (#147)
