+++ Cargo.toml
@@ -27,7 +27,6 @@ serialization_derive = { path = "serialization_derive" }
 storage = { path = "storage" }
 sync = { path = "sync" }
 test-data = { path = "test-data" }
-vdf = { path = "vdf" }
 verification = { path = "verification" }
 
 [profile.dev]
@@ -62,6 +61,5 @@ members = [
 			"storage",
 			"sync",
 			"test-data",
-			"vdf",
 			"verification",
 ]
