+++ chain/Cargo.toml
@@ -12,7 +12,6 @@ rug = "1.3.0"
 rustc-hex = "2"
 serialization = { path = "../serialization" }
 serialization_derive = { path = "../serialization_derive" }
-vdf = { path = "../vdf" }
 
 [features]
 default = [ ]
