+++ crypto/Cargo.toml
@@ -13,3 +13,6 @@ rand_core = "0.5.1"
 sha3 = "0.9.1"
 curve25519-dalek = "3.1.0"
 schnorrkel = "0.10.1"
+lazy_static = "1.3.0"
+rug = "1.3.0"
+sha2 = "0.9.1"
