--- vdf/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "vdf"
-version = "0.1.0"
-edition = "2018"
-authors = [ "RandChain https://github.com/rand-chain/" ]
-description = "Pie19 VDF (https://eprint.iacr.org/2018/627.pdf)."
-
-[dependencies]
-elapsed = "0.1.2"
-lazy_static = "1.3.0"
-rug = "1.3.0"
-serialization = { path = "../serialization" }
-serialization_derive = { path = "../serialization_derive" }
-sha2 = "0.9.1"
