--- vdf/src/config.rs
@@ -1,16 +1,0 @@
-use rug::Integer;
-use std::str::FromStr;
-
-/// RSA-2048 modulus, taken from [Wikipedia](https://en.wikipedia.org/wiki/RSA_numbers#RSA-2048).
-const RSA2048_MODULUS_DECIMAL: &str =
-  "251959084756578934940271832400483985714292821262040320277771378360436620207075955562640185258807\
-  8440691829064124951508218929855914917618450280848912007284499268739280728777673597141834727026189\
-  6375014971824691165077613379859095700097330459748808428401797429100642458691817195118746121515172\
-  6546322822168699875491824224336372590851418654620435767984233871847744479207399342365848238242811\
-  9816381501067481045166037730605620161967625613384414360383390441495263443219011465754445417842402\
-  0924616515723350778707749817125772467962926386356373289912154831438167899885040445364023527381951\
-  378636564391212010397122822120720357";
-
-lazy_static! {
-    pub static ref MODULUS: Integer = Integer::from_str(RSA2048_MODULUS_DECIMAL).unwrap();
-}
