+++ verification/Cargo.toml
@@ -17,7 +17,6 @@ serialization = { path = "../serialization" }
 sha2 = "0.9.1"
 storage = { path = "../storage" }
 time = "0.1"
-vdf = { path = "../vdf" }
 
 [dev-dependencies]
 assert_matches = "1.3.0"
