+++ verification/src/lib.rs
@@ -69,7 +69,6 @@ extern crate network;
 extern crate primitives;
 extern crate serialization as ser;
 extern crate storage;
-extern crate vdf;
 
 mod canon;
 pub mod constants;
